adopt VM cache
