+++ stylus-sdk/src/call/transfer.rs
@@ -1,14 +1,14 @@
-// Copyright 2022-2023, Offchain Labs, Inc.
+// Copyright 2022-2024, Offchain Labs, Inc.
 // For licensing, see https://github.com/OffchainLabs/stylus-sdk-rs/blob/stylus/licenses/COPYRIGHT.md
 
 use crate::call::RawCall;
 use alloc::vec::Vec;
 use alloy_primitives::{Address, U256};
 
-#[cfg(all(feature = "storage-cache", feature = "reentrant"))]
+#[cfg(feature = "reentrant")]
 use crate::storage::TopLevelStorage;
 
-#[cfg(all(feature = "storage-cache", feature = "reentrant"))]
+#[cfg(feature = "reentrant")]
 use crate::storage::Storage;
 
 /// Transfers an amount of ETH in wei to the given account.
@@ -18,7 +18,7 @@ use crate::storage::Storage;
 /// If this is not desired, the [`call`](super::call) function may be used directly.
 ///
 /// [`call`]: super::call
-#[cfg(all(feature = "storage-cache", feature = "reentrant"))]
+#[cfg(feature = "reentrant")]
 pub fn transfer_eth(
     _storage: &mut impl TopLevelStorage,
     to: Address,
@@ -43,7 +43,7 @@ pub fn transfer_eth(
 /// transfer_eth(recipient, value)?;                 // these two are equivalent
 /// call(Call::new().value(value), recipient, &[])?; // these two are equivalent
 /// ```
-#[cfg(not(all(feature = "storage-cache", feature = "reentrant")))]
+#[cfg(not(feature = "reentrant"))]
 pub fn transfer_eth(to: Address, amount: U256) -> Result<(), Vec<u8>> {
     RawCall::new_with_value(amount)
         .skip_return_data()
